+++ apps/chuni-web/src/routes/tests/components/+page.svelte
@@ -18,7 +18,7 @@
     <Rating rating={16.39} calculatedRating={16.3999} />
   </div>
 
-  <p>Calculation Error</p>
+  <p>Calculation Error (Actual 16.55, Calculated: 16.5620)</p>
 
   <div class="items-center flex gap-1 ml-4">
     <Rating rating={16.55} calculatedRating={16.562} />
@@ -55,6 +55,28 @@
       playerLevel: 169,
       playerName: "MAFUYU",
       classEmblem: 6,
+      rating: 16.39,
+      overpowerValue: 69420.44,
+      overpowerPercent: 69.69,
+      lastPlayed: new Date("2025-05-05T00:00:00Z"),
+      playCount: 999,
+    }}
+    calculatedRating={16.3999}
+  />
+  <p>I forgor to handle player reborn (Lv 100+) 💀</p>
+
+  <p>No Team?</p>
+  <Profile
+    profile={{
+      characterImage: "/mafuyu-bnw.png",
+      characterRarity: "BRONZE",
+      // teamName: "Team Name",
+      // teamEmblem: "RAINBOW",
+      honorText: "祈 -我ら神祖と共に歩む者なり-",
+      honorRarity: "GOLD",
+      playerLevel: 169,
+      playerName: "MAFUYU",
+      classEmblem: 1,
       rating: 16.39,
       overpowerValue: 69420.44,
       overpowerPercent: 69.69,